Add files via upload
Ionospheric Scintillation Morphological Analysis Using GPS-SCINDA Data at Low Latitude Ground Station
Mohammed Ayalew and Tsegaye Kassa

Abstract
This study aims to present the morphology of GPS L-band scintillations at the equatorial anomaly station Bahir Dar (11030’N, 37030’E) using GPS-SCINDA data in the descending high solar activity period between January 2014 to December 2014. In studying low-latitude scintillation, we have used millions of data recorded every minute of one year by 32 GPS satellite and it is found that intense scintillation occurred during the day time with a small frequency and very frequent occurrences of relatively moderate scintillation during the night time. In the period of observation, the variation of scintillations with local time and season are analyzed and it is found that occurrence of scintillation is minimum in summer months and maximum in equinox months with highest values observed in the months of March and September. Pre-midnight and post-midnight occurrence of scintillation is also studied and Pre-midnight scintillation was found to be maximum in equinox whereas it is minimum in winter months. Generally, it is found that most of scintillations are weak (s4<0.1) and intense scintillations with s4>0.3 are rare.


Ionospheric scintillation index and total electron characteristics are plotted by using more than five million raw data from 32 satellites. Figure 3 presents the diurnal variation of S4 from January to December 2014. The general trend shown by the surface plots indicate that the S4 value range from 0.02 to 1.14. Particularly, the smallest values are observed from 0200 to 2000 in almost the cases. The relatively high values are observed from 2000 to 0200 and ranged from 0.08 to 0.14. This general trend confirms the fact that the scintillation is significant during the night. However, it can be seen from figure three that small amount of intense scintillation occurred during the day time. The result generally exhibits an occurrence frequency of scintillation observed mainly at nighttime hours (2000-0000 extended to 0200 LT in some cases).


Based on the analyzed amplitude scintillation GPS-SCINDA data for year 2014, the following conclusions can be drawn.
•	At our location we have frequent occurrence of weak ionospheric scintillations, while few intense scintillations.
•	Intense scintillation occurred during the day time with a small frequency and relatively moderate scintillation occurrence during the night time with very frequent occurrence of scintillation.
•	The maximum numbers of amplitude scintillation events are observed in equinox and minimum in summer.
•	In summer, winter and equinox, high scintillations occurred at 19:00:00, 21:00:00 and 22:00:00 in pre-midnight local time hours respectively, while at in post mid-night period it is 04:00:00, 00:00:00 and 01:00:00 local time hours respectively.